+++ x/evm/embeds/embeds.go
@@ -43,6 +43,10 @@ var (
 	testERC20TransferWithFee []byte
 	//go:embed artifacts/contracts/TestRandom.sol/TestRandom.json
 	testRandom []byte
+	//go:embed artifacts/contracts/MKR.sol/DSToken.json
+	testMetadataBytes32 []byte
+	//go:embed artifacts/contracts/TestPrecompileSendToBankThenERC20Transfer.sol/TestPrecompileSendToBankThenERC20Transfer.json
+	testPrecompileSendToBankThenERC20Transfer []byte
 )
 
 var (
@@ -141,6 +145,17 @@ var (
 		Name:      "TestRandom.sol",
 		EmbedJSON: testRandom,
 	}
+	// SmartContract_TestBytes32Metadata is a test contract which tests contract that have bytes32 as metadata
+	SmartContract_TestBytes32Metadata = CompiledEvmContract{
+		Name:      "MKR.sol",
+		EmbedJSON: testMetadataBytes32,
+	}
+
+	// SmartContract_TestPrecompileSendToBankThenERC20Transfer is a test contract that sends to bank then calls ERC20 transfer
+	SmartContract_TestPrecompileSendToBankThenERC20Transfer = CompiledEvmContract{
+		Name:      "TestPrecompileSendToBankThenERC20Transfer.sol",
+		EmbedJSON: testPrecompileSendToBankThenERC20Transfer,
+	}
 )
 
 func init() {
@@ -158,6 +173,8 @@ func init() {
 	SmartContract_TestInfiniteRecursionERC20.MustLoad()
 	SmartContract_TestERC20TransferWithFee.MustLoad()
 	SmartContract_TestRandom.MustLoad()
+	SmartContract_TestBytes32Metadata.MustLoad()
+	SmartContract_TestPrecompileSendToBankThenERC20Transfer.MustLoad()
 }
 
 type CompiledEvmContract struct {